requirements.txt
Scrapy==1.5.1
newspaper3k==0.2.8
spacy==2.0.12
nltk==3.3
urllib3==1.25.6
sklearn==0.0
scikit-learn==0.20.2
scipy==1.1.0
numpy==1.15.0
gensim==3.6.0
matplotlib==3.0.3
pyLDAvis==2.1.2
